finish menuitem, order, paymentm, reservation and table dao

+++ src/main/java/com/example/restaurantordersystem/dao/PaymentsDAO.java
@@ -5,7 +5,7 @@ import com.example.restaurantordersystem.model.Payment;
 import java.util.List;
 
 public interface PaymentsDAO {
-    Payment findPaymentByID(long paymentID);
+    Payment findPaymentByID(int paymentID);
 
     List<Payment> findAllPayments();
